Split up query tests, separate them per operation
