+++ test/integration/Query/DatabaseTest.php
@@ -51,32 +51,4 @@ class DatabaseTest extends BaseTestCase
 
         $this->assertObStatus(['tables_dropped' => 1], $res->getData());
     }
-
-    /**
-     * @param $status
-     * @param $data
-     * @throws \Exception
-     */
-    protected function assertObStatus($status, $data)
-    {
-        $res = [];
-        $statuses = [
-            'tables_created',
-            'tables_dropped',
-            'errors',
-        ];
-        $data = new ArrayObject($data);
-
-        foreach ($statuses as $s) {
-            $status[$s] = $status[$s] ?? 0;
-        }
-
-        $data->setFlags($data::ARRAY_AS_PROPS);
-
-        foreach ($statuses as $s) {
-            $res[$s] = $data[$s] ?? 0;
-        }
-
-        $this->assertEquals($status, $res);
-    }
 }
